fix playback bug and better package structure
