+++ package.json
@@ -18,7 +18,7 @@
     "@tanstack/query-sync-storage-persister": "^4.14.5",
     "@tanstack/react-query": "^4.13.5",
     "@tanstack/react-query-persist-client": "^4.14.5",
-    "@tanstack/react-virtual": "^3.0.0-beta.18",
+    "@tanstack/react-virtual": "^3.0.0-beta.41",
     "@types/react-window": "^1.8.5",
     "@types/spotify-web-playback-sdk": "^0.1.15",
     "axios": "^1.2.6",
