+++ src/component/LikeButton.tsx
@@ -4,7 +4,7 @@ import { atom, Atom, PrimitiveAtom, useAtom, useAtomValue, useSetAtom } from 'jo
 import { nowPlayingIsLikedAtom, nowPlayingTrackIdAtom } from 'state/Atoms'
 import { useEffect, useMemo } from 'react'
 import { useQueryClient } from '@tanstack/react-query'
-import { useRemoveSavedTracksMutation, useSaveTracksMutation, useTrackLikeQuery } from './playbackSDK/hooks'
+import { useRemoveSavedTracksMutation, useSaveTracksMutation, useTrackLikeQuery } from './sdk/ClientHooks'
 import { useTransientAtom } from 'hook/useTransientAtom'
