+++ src/component/detailedReview/comment/DetailedComment.tsx
@@ -8,7 +8,7 @@ import UserAvatar, { TooltipPos } from 'component/UserAvatar'
 import { ArrowDownIcon, ArrowUpIcon, EditIcon, HazardIcon, PlayIcon, ReplyIcon, SearchIcon, TrashIcon } from 'component/Icons'
 import { useQueryClient } from '@tanstack/react-query'
 import CommentMarkdown from './CommentMarkdown'
-import { usePlayMutation } from 'component/playbackSDK/hooks'
+import { usePlayMutation } from 'component/sdk/ClientHooks'
 import { padTime } from 'util/Utils'
 import { ReviewOverview } from '../table/Helpers'
