+++ src/component/detailedReview/track/AlbumTrack.tsx
@@ -3,7 +3,7 @@ import { PrimitiveAtom } from 'jotai'
 import { RefObject, useRef } from 'react'
 import useDoubleClick from 'hook/useDoubleClick'
 import LikeButton from 'component/LikeButton'
-import { usePlayMutation } from 'component/playbackSDK/hooks'
+import { usePlayMutation } from 'component/sdk/ClientHooks'
 import { useLikeSvgStyle, useTrackColor } from './useSyncedStyles'
 import { classNames, msToTimeStr } from 'util/Utils'
 import { FireIcon } from '@heroicons/react/20/solid'
