+++ src/component/detailedReview/track/PlaylistTrack.tsx
@@ -4,7 +4,7 @@ import { RefObject, useRef } from 'react'
 import UserAvatar, { TooltipPos } from 'component/UserAvatar'
 import useDoubleClick from 'hook/useDoubleClick'
 import LikeButton from 'component/LikeButton'
-import { useAddTracksToPlaylistMutation, usePlayMutation, useReorderPlaylistTracksMutation } from 'component/playbackSDK/hooks'
+import { useAddTracksToPlaylistMutation, usePlayMutation, useReorderPlaylistTracksMutation } from 'component/sdk/ClientHooks'
 import { useLikeSvgStyle, useTrackColor } from './useSyncedStyles'
 import { classNames, msToTimeStr } from 'util/Utils'
 import { useDrag, useDrop } from 'react-dnd'
